### 2020-01-17 #2020 @whedon
[PRE REVIEW]: SeisModels.jl: A Julia package for models of the Earth's interior
**Submitting author:** @anowacki (<a href="http://orcid.org/0000-0001-7669-7383">Andy Nowacki</a>)
**Repository:** <a href="https://github.com/anowacki/SeisModels.jl" target ="_blank">https://github.com/anowacki/SeisModels.jl</a>
**Version:** v1.0.0
