### 2020-01-18 #2020 @kbarnhart
@anowacki I have a couple of clarification questions:

* Could you give me an example of using SeisModels.jl in a research application? I know the paper indicates that this package is used in research projects, but a specific example would be helpful.

* The paper makes a performance claim but I don't see any demonstration of this claim (in the paper or the docs). Could you elaborate on this claim.

>  The software has been benchmarked against independent implementations and has a significant speed advantage over similar Python software.
