### 2020-01-20 #2020 @anowacki
Thank you for the questions, @kbarnhart.

> * Could you give me an example of using SeisModels.jl in a research application? I know the paper indicates that this package is used in research projects, but a specific example would be helpful.

The software has been used by a research group here at the University of Leeds to investigate 1D models of the Earth in our group meetings.  Some of these studies are available at https://github.com/anowacki/DERG_PREM.  There is also an article in review for inclusion in an AGU monograph (Nowacki & Cottaar, 2020) which has used SeisModels.

If you feel the manuscript itself should contain examples of usage or text similar to that above, I would be happy to include that.

> * The paper makes a performance claim but I don't see any demonstration of this claim (in the paper or the docs). Could you elaborate on this claim.
>> The software has been benchmarked against independent implementations and has a significant speed advantage over similar Python software.

The benchmarking was done against a simple equivalent, and unpublished, Python code, which is available here: https://github.com/andreww/prem4derg.  I am not aware of any other open-source and freely-available software apart from that which works similarly to SeisModels in calculating bulk Earth properties, which is why concrete performance benchmarks against other implementations aren't included.

Would it be better to remove the 'comparison' and simply state some times, such as: "For a PREM-like model, computation of the surface gravity takes 5 µs, whilst calculating pressure at the centre of the body takes 12 ms."

Note that the equivalent calculations for the Python code take respectively 560 µs and 720 ms.

### References
- Nowacki, A and Cottaar, S, 2020.  Towards imaging flow at the base of the mantle with seismic, mineral physics and geodynamic constraints.  In Marquardt, H, Ballmer, M, Cottaar, S and Konter, J.  Mantle Convection and Surface Expressions.  Geophysical Monograph, American Geophysical Union, Washington, D.C., USA (in revision).
