### 2020-01-20 #2020 @kbarnhart
@anowacki thanks for the responses, they are very helpful.

A couple of comments from me in advance of review. My comments are mostly things that I anticipate would come up in review based on looking at the package and knowing the review process and criteria.

* I would consider revising the language about performance. Specifically, I would recommend describing performance in a way a reviewer can verify because one of the checklist items that reviewers will work through states:

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

* I would add instructions regarding how to test SeisModels.jl. I see that you have automatic tests set up on TravisCI and AppVeyor, but it is not clear how a reviewer might run the tests themselves.
