### 2020-01-20 9e755d1
Update JOSS paper draft
Remove abstract performance claim; add instructions on running
the tests.
