### 2020-01-20 #2020 @anowacki
Thanks for those suggestions, @kbarnhart, and for moving things forward.

I have pushed an updated version of the manuscript to the repo's master branch with concrete example timings and instructions on how to test the package (https://github.com/anowacki/SeisModels.jl/commit/9e755d1225983960d47d10699fef76a73012a85a).
