### 2020-01-20 3c9826a
Update installation instructions to reflect registration
Replace instructions to install via a URL with those to install
via the General registry.

Also make the requirement on Plots.jl in the docs explicit.
This only affects the example on the first page.
