### 2020-01-21 f83062e
Add conversions between all 1D model types
Complete the set of conversions between 1D model types by adding
methods to each constructor which take the other two model types.

Add tests and documentation for these.
