+++ src/conversion.jl
@@ -40,3 +40,142 @@ function LinearLayeredModel(m::PREMPolyModel, spacing=20)
         isanisotropic(m), Vph, Vpv, Vsh, Vsv, Eta,
         hasattenuation(m), qμ, qκ)
 end
+
+"Set of fields we copy to the LinearLayeredModel struct"
+const LinearLayeredModel_CONVERSION_FIELDS =
+    (x for x in fieldnames(LinearLayeredModel)
+     if x ∉ (:a, :n, :r, :aniso, :attenuation))
+
+"""
+    LinearLayeredModel(m::SteppedLayeredModel)
+
+Convert the model `m` into a `LinearLayeredModel`, with radial
+knots spaced at the bottom and top of each layer in `m`.
+"""
+function LinearLayeredModel(m::SteppedLayeredModel)
+    R = surface_radius(m)
+    d = Dict{Symbol,Vector{Float64}}()
+    radii = Iterators.flatten((i == 1 ? 0.0 : m.r[i-1], m.r[i]) for i in eachindex(m.r))
+    for field in LinearLayeredModel_CONVERSION_FIELDS
+        val = getfield(m, field)
+        isempty(val) && continue
+        d[field] = repeat(val, inner=2)
+    end
+    d[:r] = collect(radii)
+    LinearLayeredModel(; d...)
+end
+
+"Set of fields we copy to the PREMPolyModel struct"
+const PREMPolyModel_CONVERSION_FIELDS = (x for x in fieldnames(PREMPolyModel)
+                                         if x ∉ (:r, :a, :n, :fref, :aniso, :attenuation))
+
+"""
+    PREMPolyModel(m::SteppedLayeredModel, order=0; fref=1.0)
+
+Convert the model `m` into a `PREMPolyModel`, with maximum polynomial
+`order`.  The reference frequency `fref` in Hz can be given.
+"""
+function PREMPolyModel(m::SteppedLayeredModel, order=0; fref=1.0)
+    order < 0 &&
+        throw(ArgumentError("cannot construct polynomials with negative order"))
+    nlayers = length(m.r)
+    d = Dict{Symbol,Any}()
+    for field in PREMPolyModel_CONVERSION_FIELDS
+        val = getfield(m, field)
+        isempty(val) && continue
+        if val isa AbstractArray && order > 0
+            arr = fill(0.0, order + 1, nlayers)
+            arr[1,:] .= val
+            d[field] = arr
+        else
+            d[field] = val
+        end
+    end
+    d[:fref] = fref
+    d[:r] = m.r
+    PREMPolyModel(; d...)
+end
+
+"""
+    PREMPolyModel(m::LinearLayeredModel, order=1; fref=1.0)
+
+Convert the model `m` into a `PREMPolyModel`, with maximum polynomial
+`order`.  The reference frequency `fref` in Hz can be given.
+"""
+function PREMPolyModel(m::LinearLayeredModel, order=1; fref=1.0)
+    order < 1 &&
+        throw(ArgumentError("minimum polynomial order is 1 when " *
+                            "converting `LinearLayeredModel`s"))
+    d = Dict{Symbol,Any}()
+    R = surface_radius(m)
+    # Indices of nodes on top of finite layers (i.e., not discontinuity layers)
+    layer_inds = [i for i in 2:length(m.r) if m.r[i] > m.r[i-1]]
+    nlayers = length(layer_inds)
+    for field in PREMPolyModel_CONVERSION_FIELDS
+        val = getfield(m, field)
+        isempty(val) && continue
+        if val isa AbstractArray
+            d[field] = fill(0.0, order + 1, nlayers)
+            for (i, ind) in enumerate(layer_inds)
[…]
