### 2020-01-21 #2020 @daniellivingston
Hi @kbarnhart , this is right up my alley. I'd be happy to review.
