### 2020-01-21 75ef3ec
Bump package version to v1.1.0
