### 2020-01-22 #2020 @kbarnhart
:wave: @lheagy @fhorrobin a friendly reminder to consider reviewing this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. Once there are sufficient reviewers (usually 2-3), I will open a new issue where the review will take place.

If you are not able, please let me know if you have recommendations for other reviewers. Please recommend without mentioning a handle (e.g., you would refer to me as @ kbarnhart instead of @kbarnhart). This way we attempt to reduce the number of people who are subscribed to this thread.
