### 2020-01-27 #2043 @whedon
[REVIEW]: SeisModels.jl: A Julia package for models of the Earth's interior
**Submitting author:** @anowacki (<a href="http://orcid.org/0000-0001-7669-7383">Andy Nowacki</a>)
**Repository:** <a href="https://github.com/anowacki/SeisModels.jl" target ="_blank">https://github.com/anowacki/SeisModels.jl</a>
**Version:** v1.1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.11993313" target="_blank">10.6084/m9.figshare.11993313</a>

@daniellivingston, @fhorrobin, & @joa-quim please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
