### 2020-01-27 #2043 @kbarnhart
:wave: @anowacki, @daniellivingston, @fhorrobin this is the review thread for the paper. All of our communications will happen here from now on.

All reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `https://github.com/openjournals/joss-reviews/issues/2043` so that a link is created to this thread. This helps create a record of the changes in the repository associated with this review. It also helps me keep track of open/closed issues that remain.

Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
