### 2020-02-03 #2020 @kbarnhart
@leouieda thanks for the recommendations.

:wave: @seisman @joa-quim would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This is a pre-review issue that I use to coordinate reviewers. If you are able, I will add you to the review issue at openjournals/joss-reviews/issues/2043.
