### 2020-02-13 #2043 @daniellivingston
@anowacki Great package. Does what it claims and does it very well and very elegantly. A minor nitpick and probably a personal taste issue; feel free to ignore.

You write that "[d]ocumentation is available at a dedicated website" without explicitly spelling out the URL. I understand the URL is relatively long (https://anowacki.github.io/SeisModels.jl/stable/), and that embedded hyperlinks mean that spelling out the full address is redundant, but I would still like to see the full path in the document. However, this is a minor thing and perhaps up to personal taste, so I will approve this paper regardless. Just something to think about.
