### 2020-02-13 #2043 @daniellivingston
Also, in the documentation (https://anowacki.github.io/SeisModels.jl/stable/#How-to-install-1), one finds this:

> ### How to install
> SeisModels.jl can be added to your Julia environment like so:
>
> ```
> julia> import Pkg; pkg"add https://github.com/anowacki/SeisModels.jl"
> ```

This should be changed to match your repo README:

> SeisModels.jl can be added to your Julia install like so:
>
> ```
> julia> import Pkg; Pkg.add("SeisModels")
> ```
