### 2020-02-13 #2043 @daniellivingston
One of the review guidelines states:

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Insofar as I can tell, discussions of third-party contributions are limited to this line in your README:

> Currently, only three kinds of one-dimensional models are supported, but all model parameterisations and models are acceptable for inclusion. Contributions are welcome.

You may want to add a CONTRIBUTION.md document to the repo root, or just a small paragraph in the README expanding on your contribution process. You can find an over-the-top example of this here: https://reactjs.org/docs/how-to-contribute.html
