### 2020-02-13 #2043 @anowacki
I have set up a branch [joss-review](https://github.com/anowacki/SeisModels.jl/tree/joss-review) in which I propose to make the changes before merging them in finally.

@daniellivingston—thank you very much for these helpful suggestions.

### Docs website URL
> You write that "[d]ocumentation is available at a dedicated website" without explicitly spelling out the URL. […]  I would still like to see the full path in the document.

I'm very happy to go along with this and I have updated the manuscript accordingly.

### Contribution guidelines
> You may want to add a CONTRIBUTION.md document to the repo root, or just a small paragraph in the README expanding on your contribution process.

Thank you for helping me realise this—I had not quite understood the guidelines here, but do now on re-reading them.

I have added a CONTRIBUTING.md file (this filename seems much more common, at least on GirHub and Gitlab repos) to the repo root which hopefully explains things in appropriate detail.

### Docs installation instructions
> This should be changed to match your repo README:

Thank you for spotting this.  I did update the docs a little while ago (in [3c9826ac](https://github.com/anowacki/SeisModels.jl/commit/3c9826acc03a38ec86e1a4a932031db7f135e7ed)), but it transpires that the docs stopped building after an [earlier commit](https://github.com/anowacki/SeisModels.jl/commit/dbb5530e523eb890226c5e221a41b4fe135397fb).  I can see that a number of new functions are in fact missing.  I will have to investigate this further to work out what has happened.
