### 2020-02-13 f788c51
Docs: Try to fix Travis config to build docs again
Docs have not been built since dbb5530e.  It appears that Travis
have changed the syntax of .travis.yml:

    https://discourse.julialang.org/t/travis-ci-configuration-breaking-changes/34268

Update .travis.yml to try and rectify this.
