### 2020-02-13 #2043 @anowacki
> ### Docs installation instructions
>> This should be changed to match your repo README:
>
> Thank you for spotting this. I did update the docs a little while ago (in 3c9826ac), but it transpires that the docs stopped building after an earlier commit. I can see that a number of new functions are in fact missing. I will have to investigate this further to work out what has happened.

Docs are now [back up](https://anowacki.github.io/SeisModels.jl/dev/).  (The issue was [with Travis configuration](https://github.com/anowacki/SeisModels.jl/commit/f788c51a60c00000e422b2d53bc58925085cd744).)
