### 2020-02-14 6ae6de1
Add STW105 reference model of Kustowski et al. (2008)
STW105 is the reference model for S362ANI and S362WMANI.
It takes Q from QL6 and matches PREM below the transition zone,
but is different above that.

The original is parameterised by raidal splines; this version
is taken from IRIS and is a LinearLayeredModel.
