+++ src/Earth/stw105.jl
@@ -1,0 +1,991 @@
+"""
+# `STW105`
+
+1-D layered model of the Earth, containing radial anisotropy,
+density and attenuation.
+
+Attenuation is taken from QL6 (Durek and Ekström, 1996).
+
+This version was obtained from:
+    https://ds.iris.edu/ds/products/emc-stw105/
+The original model is parameterised by radial splines.
+
+## Reference
+
+Kustowski, B., Ekström, G., Dziewoński, A., 2008.
+Anisotropic shear-wave velocity structure of the Earth’s mantle: A global model.
+J Geophys Res-Sol Ea 113, B06306.
+https://doi.org/10.1029/2007JB005169
+"""
+const STW105 = LinearLayeredModel(
+    r = [0., 6824., 13648., 20472., 27296., 34120., 40944., 
+         47768., 54592., 61416., 68240., 75064., 81888., 88712., 
+         95536., 102360., 109184., 116007., 122831., 129655., 136479., 
+         143303., 150127., 156951., 163775., 170599., 177423., 184247., 
+         191071., 197895., 204719., 211543., 218367., 225191., 232015., 
+         238839., 245663., 252487., 259311., 266135., 272959., 279783., 
+         286607., 293431., 300255., 307079., 313903., 320727., 327551., 
+         334375., 341199., 348022., 354846., 361670., 368494., 375318., 
+         382142., 388966., 395790., 402614., 409438., 416262., 423086., 
+         429910., 436734., 443558., 450382., 457206., 464030., 470854., 
+         477678., 484502., 491326., 498150., 504974., 511798., 518622., 
+         525446., 532270., 539094., 545918., 552742., 559566., 566390., 
+         573214., 580037., 586861., 593685., 600509., 607333., 614157., 
+         620981., 627805., 634629., 641453., 648277., 655101., 661925., 
+         668749., 675573., 682397., 689221., 696045., 702869., 709693., 
+         716517., 723341., 730165., 736989., 743813., 750637., 757461., 
+         764285., 771109., 777933., 784757., 791581., 798405., 805229., 
+         812052., 818876., 825700., 832524., 839348., 846172., 852996., 
+         859820., 866644., 873468., 880292., 887116., 893940., 900764., 
+         907588., 914412., 921236., 928060., 934884., 941708., 948532., 
+         955356., 962180., 969004., 975828., 982652., 989476., 996300., 
+         1003124., 1009948., 1016772., 1023596., 1030420., 1037244., 1044067., 
+         1050891., 1057715., 1064539., 1071363., 1078187., 1085011., 1091835., 
+         1098659., 1105483., 1112307., 1119131., 1125955., 1132779., 1139603., 
+         1146427., 1153251., 1160075., 1166899., 1173723., 1180547., 1187371., 
+         1194195., 1201019., 1207843., 1214667., 1221491., 1221491., 1234250., 
+         1247010., 1259770., 1272530., 1285289., 1298049., 1310809., 1323568., 
+         1336328., 1349088., 1361847., 1374607., 1387367., 1400127., 1412886., 
+         1425646., 1438406., 1451165., 1463925., 1476685., 1489444., 1502204., 
[…]
