+++ docs/src/references.md
@@ -1,5 +1,8 @@
 # References
 
+- Durek, J.J., Ekström, G., 1996.
+  A radial model of anelasticity consistent with long-period surface-wave attenuation.
+  _Bull Seism Soc Am_ 86, 144–158.
 - Dziewoński, A., Anderson, D., 1981. Preliminary reference Earth model.
   _Phys. Earth Planet. Inter._ 25, 297–356. doi:[10.1016/0031-9201(81)90046-7](https://doi.org/10.1016/0031-9201(81)90046-7)
 - Kennett  B.L.N.  Engdahl  E.  1991. Traveltimes for global earthquake
@@ -8,6 +11,10 @@
 - Kennett, B.L.N., Engdahl, E., Buland, R., 1995. Constraints on seismic
   velocities in the Earth from travel-times.
   _Geophys J Int_ 122, 108–124. doi:[10.1111/j.1365-246X.1995.tb03540.x](https://doi.org/10.1111/j.1365-246X.1995.tb03540.x)
+- Kustowski, B., Ekström, G., Dziewoński, A., 2008.
+  Anisotropic shear-wave velocity structure of the Earth’s mantle: A global model.
+  _J Geophys Res-Sol Ea_ 113, B06306.
+  doi:[10.1029/2007JB005169](https://doi.org/10.1029/2007JB005169)
 - Weber, R.C.,, Lin, P.-Y.,, Garnero, E.J., Williams, Q., Lognonné, P., 2011.
   Seismic Detection of the Lunar Core. _Science_, 331, 309–312.doi:[10.1126/science.1199375](https://doi.org/10.1126/science.1199375)
 
