+++ docs/src/inbuilt_models.md
@@ -1,12 +1,17 @@
 # Inbuilt models
 
 Most users of SeisModels will probably want to simply calculate
-properties for one of the inbuilt models of the Earth and moon.
-They are:
+properties for one of the inbuilt models.  These are listed below:
 
+## Earth models
 ```@docs
 AK135
 IASP91
 PREM
+STW105
+```
+
+## Moon models
+```@docs
 MOON_WEBER_2011
 ```
