### 2020-02-14 #2043 @kbarnhart
Thanks to @daniellivingston for completing your review and for @anowacki for addressing the raised issues. 🎉

@fhorrobin and @joa-quim please work to complete your reviews in the next week or so (or provide a time estimate here).

Thanks all for participating in the JOSS review process.
