### 2020-02-16 #2043 @kbarnhart
Yup. You should be able to check the checkboxes under "Review checklist for @joa-quim". I'm not entirely certain how to fix this, so I'm tagging @arfon who should be able to provide guidance.
