### 2020-02-16 #2043 @joa-quim
Thanks. And I'll complete the review this week (I've already installed the package and run some examples).
