### 2020-02-24 #2043 @kbarnhart
@joa-quim I wanted to check in and see if you were able to check the boxes or if they are still not editable for you.
