### 2020-02-24 #2043 @kbarnhart
@arfon can you see anything about this issue that explains why @joa-quim can't check the boxes? I used whedon to assign him and he is now listed under assigned people. Thanks!
