### 2020-02-25 #2043 @arfon
@joa-quim - you need to accept the invite to the repository here: https://github.com/openjournals/joss-reviews/invitations  to be able to edit the check boxes (it should be at the top of the page). When you click the 'accept', please make sure you're logged in with your @joa-quim account.
