+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"references/#References-1","page":"References","title":"References","text":"","category":"section"},{"location":"references/#","page":"References","title":"References","text":"Dziewoński, A., Anderson, D., 1981. Preliminary reference Earth model. Phys. Earth Planet. Inter. 25, 297–356. doi:10.1016/0031-9201(81)90046-7\nKennett  B.L.N.  Engdahl  E.  1991. Traveltimes for global earthquake location and phase identification. Geophys J Int 105  429–465. doi:10.1111/j.1365-246X.1991.tb06724.x\nKennett, B.L.N., Engdahl, E., Buland, R., 1995. Constraints on seismic velocities in the Earth from travel-times. Geophys J Int 122, 108–124. doi:10.1111/j.1365-246X.1995.tb03540.x\nWeber, R.C.,, Lin, P.-Y.,, Garnero, E.J., Williams, Q., Lognonné, P., 2011. Seismic Detection of the Lunar Core. Science, 331, 309–312.doi:10.1126/science.1199375","category":"page"},{"location":"examples/#Examples-1","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"examples/#Preamble-1","page":"Examples","title":"Preamble","text":"","category":"section"},{"location":"examples/#","page":"Examples","title":"Examples","text":"These commands all assume that you have done","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"using SeisModels","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"first.","category":"page"},{"location":"examples/#Basic-properties-1","page":"Examples","title":"Basic properties","text":"","category":"section"},{"location":"examples/#","page":"Examples","title":"Examples","text":"To compute the P-wave velocity at the base of the Earth’s mantle (2750 km depth) in the iasp91, ak135 and PREM models, and compare them all to PREM, you could do","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"vels = vp.((IASP91, AK135, PREM), 2750, depth=true)\n100 .* (vels .- vels[end]) ./ vels[end]","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"(So they all agree to less than a quarter of a percent!)","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"To retrieve a density profile across the inner core boundary in PREM:","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"r = 1200:10:1250;\ndensity.(PREM, r)","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"To get the planetary radius in km of the moon used in Weber et al.’s (2011) model:","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"surface_radius(MOON_WEBER_2011)","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"The shear modulus in the transition zone:","category":"page"},{"location":"examples/#","page":"Examples","title":"Examples","text":"shear_modulus(AK135, 600, depth=true)","category":"page"},{"location":"examples/#Derived-properties-1","page":"Examples","title":"Derived properties","text":"","category":"section"},{"location":"examples/#","page":"Examples","title":"Examples","text":"If a model includes density, then things like gravity, moment of inertia, and so on, can be easily calculated.","category":"page"},{"location":"examples/#","page":"Examples","title":
[…]
