### 2020-02-25 #2043 @joa-quim
Hmm, something strange is still on. When I try to accept the invitation I get a
```Ooops 500. Looks like something went wrong!```
 error from Github.

Anyway, my comment on this paper is that it is elegantly written and does what it describes,

I would only suggest that the installing instructions be changed to the shorter (and working)

```
]add SeisModels
```
instead of the longer version on the Github page
```
import Pkg; Pkg.add("SeisModels")
```
or the even longer in the documentation
```
import Pkg; pkg"add https://github.com/anowacki/SeisModels.jl"
```

Given that I'm still not able to check the boxes above, I reproduced the check boxes list bellow with the corresponding checks. Hope this is good enough as a workaround.

# --------------------

- Conflict of interest

- [x] I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

- Code of Conduct

- [x] I confirm that I read and will adhere to the JOSS code of conduct.

- General checks

- [x] **Repository:** Is the source code for this software available at the repository url?
- [x] **License:** Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

- Functionality

- Documentation

- Software paper
