### 2020-02-25 #2043 @kbarnhart
Thanks for your review @joa-quim! Sorry about the issues with check box permissions. I think your workaround is just fine.

I'll let @anowacki consider/address the small change you recommend in the main repository.

@fhorrobin would it be possible to get an update on when you expect to complete your review?
