### 2020-02-27 #2043 @anowacki
Thanks, @joa-quim for the comments!

> I would only suggest that the installing instructions be changed to the shorter (and working)
>     ```]add SeisModels```
> instead of the longer version on the Github page
>     ```import Pkg; Pkg.add("SeisModels")```

My motivation for the latter command is that new Julia users can simply copy and paste the full command into a fresh REPL session without needing to worry about `pkg` mode, but I'd be happy to change this is you feel it is definitely less friendly.

> or the even longer in the documentation
>     ```import Pkg; pkg"add https://github.com/anowacki/SeisModels.jl"```

When the package wasn't registered, this was needed, but no longer as I'm sure you realise.  This part of the docs is a leftover from before [I realised the docs weren't updating](https://github.com/openjournals/joss-reviews/issues/2043#issuecomment-585993988) and the [latest docs](https://anowacki.github.io/SeisModels.jl/dev/) do not add using the GirHub URL.
