### 2020-02-27 #2043 @joa-quim
@anowacki This was just my opinion on what looks more friendly to new users. More experimented users should know that already. But notice that one can do a copy past of ``]add SeisModels`` string into a fresh REPL and it works fine. I mean, the REPL passes automatically into the *pkg* mode.
