### 2020-02-27 #2043 @anowacki
> But notice that one can do a copy past of `]add SeisModels` string into a fresh REPL and it works fine. I mean, the REPL passes automatically into the _pkg_ mode.

This doesn't actually work for me.
