### 2020-02-27 #2043 @joa-quim
!! I'm on Windows and I was assuming that it would work in other OSs as well.
Anyway, this is minor point. Do as you feel is better.
