### 2020-03-02 #2043 @anowacki
Thanks for you suggestion, @joa-quim—just for the avoidance of doubt in the review process, I'm stating that I have decided to keep the docs as they are in this regard.
