### 2020-03-03 #2043 @kbarnhart
@anowacki Thanks for noting your response @joa-quim's comment here for posterity.

@fhorrobin would it be possible to get an update on when you expect to complete your review?

@joa-quim and @daniellivingston thank you for completing your reviews.

A quick additional note. I accidentally turned on the JOSS out of office responder and have not figured out how to turn it off... So if you @ me, you will see an auto response that says I'm away until March 20th. This is not true, I'm not away... 😄
