### 2020-03-10 #2043 @kbarnhart
Thanks to @daniellivingston and @joa-quim for completing your reviews.

@fhorrobin a friendly reminder to provide an update on when you expect to complete your review?  If I don't hear back in the next day or two, I will likely decide to move forward with the review process (as JOSS only requires two reviews).
