### 2020-03-11 #2043 @kbarnhart
I'm going to move this review forward.

Now that the two reviews have been completed, @daniellivingston and @joa-quim could you please reply and confirm your review recommendation.

At that point @anowacki could you please point me towards the current version of the paper (it is currently on a branch I think). I will provide some final, likely minor, editorial comments on the paper.

Once those are merged into the main repository branch, we will move to the final stages of the JOSS process (double checking reference DOIs, final paper compilation, setting archive, and version). @anowacki I'll ping you for what I need when we get there.

@fhorrobin thanks for being willing to undertake a review. It seems like things didn't work out with your schedule, no worries. I hope you'll consider being part of a JOSS review in the future. I'm going to remove you from this review now. Feel free to follow up with me here or via email if you'd like.
