### 2020-03-13 #2043 @kbarnhart
@anowacki is this the most up to date version of the publication? I may have a couple of minor editorial comments on the paper---just want to make sure I'm looking at the right thing.
