### 2020-03-16 #2043 @anowacki
> @anowacki is this the most up to date version of the publication? I may have a couple of minor editorial comments on the paper---just want to make sure I'm looking at the right thing.

Hi @kbarnhart—yes, this version is the current one, so please do let me know of any comments you have, the current situation permitting of course.
