### 2020-03-16 #2043 @kbarnhart
@anowacki I only have a couple of comments. They are meant to improve readability and accessibility for a non-specialist audience.

- [ ] In paragraph 2: Define abbreviations: I think this only means writing out PREM as Preliminary Reference Earth Model. If you are aware of better written out names of the other Earth models, I'd recommend using them, but I suspect they will stay as AK135 and iasp91. I think P-wave and S-wave can stay as is.

- [ ] In paragraph 4: "appropriate formats" is rather unspecific. Could you revise to state which formats are supported.

Once these changes are in the paper, please do the following:

- [ ] Make a tagged release of your software
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (title and author list exactly the same as on the JOSS paper).
- [ ] Comment here with:
    - The version tag of the archived version
    - the DOI of the archived version

I with then move forward with accepting the submission (I will do some final metadata updating and then pass the submission on to the managing editor in chief who will handle final acceptance and publication).
