### 2020-03-16 48e00c4
Update JOSS paper with editorial corrections.
Update the JOSS paper to address the comments here:

    https://github.com/openjournals/joss-reviews/issues/2043#issuecomment-599605209
