### 2020-03-16 3c05cf8
Bump version to v1.2.0
