+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.2",
   "v1.0",
   "v0.3",
   "dev",
