### 2020-03-17 #2043 @anowacki
@kbarnhart I have archived v1.1.0 of SeisModels.jl at Figshare with doi 10.6084/m9.figshare.11993313.
