### 2020-03-17 #2043 @kbarnhart
Nice work @anowacki! Thanks to @daniellivingston and @joa-quim for your reviews.

I've now recommended that this be accepted and handed it off to the JOSS Editors in Chief who handle final processing.
