### 2020-03-17 #2043 @arfon
@daniellivingston, @joa-quim - many thanks for your reviews and to @kbarnhart for editing this submission ✨

@anowacki - your paper is now published in JOSS :zap::rocket::boom:
